Correctly reflect unknown lag in graph
Since we are missing data from before 2020-04-03 in our dataset we have
no way of knowing when most of the deaths before that date were reported
and tht should be reflected in the graph somehow.

Before this commit the deaths represented at an unknown date were
graphed with a lag of zero. This commit changes that to a lag of NA
which removes a potentially missleading thing in the graph.

As a side effect this commit reverses the colors of the graph, but it
should be just as readable.